career-launchpad/rag-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
chromadb==0.6.3
sentence-transformers==3.4.1
openai==1.60.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.5
pypdf==5.1.0
python-multipart==0.0.18
